fix: set bus type from VM preference
Currently the temporary VM is created with
hardcoded value of bus type, which is virtio,
and is not configurable.

This modification will allow to set the VM's
bus type from the VM preference, and
also an option to choose which kind of the
instance type or preference to use, either
namespaced one or from the cluster.

Signed-off-by: Ben Oukhanov <boukhanov@redhat.com>